feat(ts): move other declarations to d.ts files
