+++ typings/react-intl.d.ts
@@ -1,0 +1,18 @@
+declare module 'react-intl' {
+  import { StatelessComponent } from 'react'
+
+  export const injectIntl
+
+  export const FormattedMessage: StatelessComponent<{
+    className?: string
+    defaultMessage: string
+    id: string
+    values?: any
+  }>
+
+  export const FormattedHTMLMessage: StatelessComponent<{
+    defaultMessage: string
+    id: string
+    values?: any
+  }>
+}
